+++ src/GameTrilha.GameDomain/Entities/PiecesAmount.cs
@@ -2,4 +2,4 @@
 
 namespace GameTrilha.GameDomain.Entities;
 
-public record PiecesAmount(Color Color, int Amount);
+public record PiecesAmount(Color Color, byte Amount);
